// src/app/core/models/movie.model.ts
/**
 * Movie Collection interface
 */
export interface MovieCollection {
  id: number;
  name: string;
  poster_path: string;
  backdrop_path: string;
}

/**
 * Genre interface
 */
export interface Genre {
  id: number;
  name: string;
}

/**
 * Production Company interface
 */
export interface ProductionCompany {
  id: number;
  logo_path: string;
  name: string;
  origin_country: string;
}

/**
 * Production Country interface
 */
export interface ProductionCountry {
  iso_3166_1: string;
  name: string;
}

/**
 * Spoken Language interface
 */
export interface SpokenLanguage {
  english_name: string;
  iso_639_1: string;
  name: string;
}

/**
 * Movie interface representing movie data from API
 */
export interface Movie {
  adult: boolean;
  backdrop_path: string;
  belongs_to_collection: MovieCollection | null;
  budget: number;
  genres: Genre[];
  homepage: string;
  id: number;
  imdb_id: string;
  origin_country: string[];
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: ProductionCompany[];
  production_countries: ProductionCountry[];
  release_date: string;
  revenue: number;
  runtime: number;
  spoken_languages: SpokenLanguage[];
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}